Move cys ai tests to test folder
