+++ plugins/woocommerce-admin/client/customize-store/design-with-ai/prompts/test/colorChoices.test.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { defaultColorPalette } from '.';
+import { defaultColorPalette } from '..';
 
 describe( 'colorPairing.responseValidation', () => {
 	it( 'should validate a correct color palette', () => {
