+++ plugins/woocommerce-admin/client/customize-store/design-with-ai/prompts/test/fontPairings.test.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { fontChoiceValidator } from '.';
+import { fontChoiceValidator } from '..';
 
 describe( 'fontChoiceValidator', () => {
 	it( 'should validate when font choice is part of the allowed list', () => {
